+++ Enderlook.Pools/src/ExactLengthArrayPool.cs
@@ -1,6 +1,7 @@
 ﻿using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Threading;
 
@@ -10,9 +11,34 @@ namespace Enderlook.Pools;
 /// Represent a pool of arrays of exact length.
 /// </summary>
 /// <typeparam name="T">Type of object to pool.</typeparam>
-public abstract class ExactLengthArrayPool<T> : ArrayPool<T>
+public abstract class ExactLengthArrayPool<T> : ArrayPool<T>, IReturnable<T[]>
 {
     private protected Dictionary<int, ObjectPool<T[]>>? adapters;
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    void IReturnable<T[]>.Return(T[] element, bool state) => Return(element, state);
+
+    /// <summary>
+    /// Rent an element from the pool.<br/>
+    /// If the pool is empty, instantiate a new element.<br/>
+    /// Implementors of this interface can choose how elements are instantiated and initialized, or throw if instantiation of new elements is not supported.
+    /// </summary>
+    /// <param name="minimumLength">The minimum length of the array needed.</param>
+    /// <returns>A contained of the rented element. When this container is disposed, the object will be returned to the pool.</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public Lease<T[]> RentLease(int minimumLength) => new(this, false, Rent(minimumLength));
+
+    /// <summary>
+    /// Rent an element from the pool.<br/>
+    /// If the pool is empty, instantiate a new element.<br/>
+    /// Implementors of this interface can choose how elements are instantiated and initialized, or throw if instantiation of new elements is not supported.
+    /// </summary>
+    /// <param name="minimumLength">The minimum length of the array needed.</param>
+    /// <param name="clearOnReturn"> If <see langword="true"/> and if the pool will store the buffer to enable subsequent reuse, will clear the array of its contents so that a subsequent consumer will not see the previous consumer's content.<br/>
+    /// If <see langword="false"/> or if the pool will release the buffer, the array's contents are left unchanged.</param>
+    /// <returns>A contained of the rented element. When this container is disposed, the object will be returned to the pool.</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public Lease<T[]> RentLease(int minimumLength, bool clearOnReturn) => new(this, clearOnReturn, Rent(minimumLength));
 
     /// <summary>
     /// Retrieves a shared <see cref="ExactLengthArrayPool{T}"/> instance.<br/>